+++ week4/assignment/로그인 & 회원가입 👤/src/components/login/Login.jsx
@@ -5,7 +5,7 @@ import axios from "axios";
 
 import { INPUT } from "../../constants/input";
 
-import Toast from "./Toast";
+import Toast from "../common/Toast";
 
 const Login = ({ info }) => {
   const requestData = {
